Добавил селекторы в selectors.js
